feat : sign up

+++ FashionToday_Back/src/main/java/hamburger/fashiontoday/domain/member/MemberRepository.java
@@ -2,6 +2,6 @@ package hamburger.fashiontoday.domain.member;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface MemberRepository extends CrudRepository<Member,Long> {
-    public Member findByMId(Long mId);
+public interface MemberRepository extends CrudRepository<Member, Integer> {
+    public Member findByMId(int mId);
 }
